Renaming util components to be more descriptive
