+++ .storybook/styles.scss
@@ -31,8 +31,6 @@ pre {
 
 .example--code {
   padding: 10px;
-  // background: #111;
-  // border-radius: 3px;
   color: #000;
   font-family: monospace;
   font-size: 0.9em;
